backend/requirements.txt
# Core Django
Django==4.2

# Django REST Framework
djangorestframework==3.14.0

# PostgreSQL driver
psycopg2-binary==2.9.6

# CORS handling
django-cors-headers==3.14.0

# Image handling
Pillow==10.1.0

# JWT Authentication
djangorestframework-simplejwt==5.4.0

# OpenAI
openai==1.17.0

# HTTP Requests
requests==2.31.0

dj-database-url
gunicorn
whitenoise
python-dotenv